* lock.c (set_lockers_name): Use %lu in format and cast st_uid   field to unsigned long.
* import.c (import): Use RCS_check_kflag() to check -k options.
  (keyword_usage, str2expmode, strn2expmode, expand_names):
  Removed.
* rcs.c (RCS_check_kflag): Added keyword_usage array from import.c
  for more descriptive error messages.

CVS-ID: src/ChangeLog 1.66
CVS-ID: src/import.c 1.8
CVS-ID: src/lock.c 1.7
CVS-ID: src/rcs.c 1.9

+++ src/lock.c
@@ -428,7 +428,7 @@ set_lockers_name (statp)
 	(void) strcpy (lockers_name, pw->pw_name);
     }
     else
-	(void) sprintf (lockers_name, "uid%d", statp->st_uid);
+	(void) sprintf (lockers_name, "uid%lu", (unsigned long) statp->st_uid);
 }
 
 /*
